+++ src/modules/orders/dto/getOrders.args.ts
@@ -1,4 +1,4 @@
-import { ArgsType } from '@nestjs/graphql';
+import { ArgsType, Field } from '@nestjs/graphql';
 import { PaginationArgs } from '@pagination/pagination.args';
 
 import { Order } from '../models/order.model';
@@ -6,6 +6,8 @@ import { GeneratePaginationOutput } from '@pagination/pagination.output';
 
 @ArgsType()
 export class GetOrdersWithPaginationArgs extends PaginationArgs {
-};
+  @Field(() => String, { nullable: true })
+  email?: string;
+}
 
 export const OrderOutput = GeneratePaginationOutput<Order>(Order);